Refactor FuncParam class
